Add type cast to `readline(2)` return value.

+++ mrbgems/mruby-io/src/file.c
@@ -370,7 +370,7 @@ mrb_file_s_readlink(mrb_state *mrb, mrb_value klass) {
   mrb_get_args(mrb, "z", &path);
 
   buf = (char *)mrb_malloc(mrb, bufsize);
-  while ((rc = readlink(path, buf, bufsize)) == bufsize && rc != -1) {
+  while ((rc = readlink(path, buf, bufsize)) == (ssize_t)bufsize && rc != -1) {
     bufsize *= 2;
     buf = (char *)mrb_realloc(mrb, buf, bufsize);
   }
